### 2017-11-01 #446 @whedon
[PRE REVIEW]: PFEIFER: Preprocessing Framework of Electrograms Intermittently Fiducialized from Experimental Recordings
**Submitting author:** @bzenger (<a href="http://orcid.org/0000-0002-0039-9184">Brian Zenger</a>)
**Repository:** <a href="https://github.com/SCIInstitute/PFEIFER" target ="_blank">https://github.com/SCIInstitute/PFEIFER</a>
**Version:** v1.0.0
