### 2017-11-01 1c3c1c7
fixed fiducial shift error in autofiducializing
