+++ SOURCE/GUI_Scripts/fidsDisplay.m
@@ -525,154 +525,155 @@ function SetupDisplay(handle)
 %         - sets up some start values for xlim, ylim, sets up axes and slider handles
 %         - makes the FD.SIGNAL values,   (RMS and scaling of potvals)
 
-    pointer = handle.Pointer;
-    handle.Pointer = 'watch';
-    global TS SCRIPTDATA FIDSDISPLAY;
-    
-    tsindex = SCRIPTDATA.CURRENTTS;
-    
-    numframes = size(TS{tsindex}.potvals,2);
-    FIDSDISPLAY.TIME = [1:numframes]*(1/SCRIPTDATA.SAMPLEFREQ);
-    FIDSDISPLAY.XLIM = [1 numframes]*(1/SCRIPTDATA.SAMPLEFREQ);
+pointer = handle.Pointer;
+handle.Pointer = 'watch';
+global TS SCRIPTDATA FIDSDISPLAY;
+
+tsindex = SCRIPTDATA.CURRENTTS;
+
+numframes = size(TS{tsindex}.potvals,2);
+FIDSDISPLAY.TIME = [1:numframes]*(1/SCRIPTDATA.SAMPLEFREQ);
+FIDSDISPLAY.XLIM = [1 numframes]*(1/SCRIPTDATA.SAMPLEFREQ);
 
 %    if isempty(FIDSDISPLAY.XWIN);
-        FIDSDISPLAY.XWIN = [median([0 FIDSDISPLAY.XLIM]) median([3000/SCRIPTDATA.SAMPLEFREQ FIDSDISPLAY.XLIM])];
-        %    else
-        %FIDSDISPLAY.XWIN = [median([FIDSDISPLAY.XWIN(1) FIDSDISPLAY.XLIM]) median([FIDSDISPLAY.XWIN(2) FIDSDISPLAY.XLIM])];
-        %end
-    
-    FIDSDISPLAY.AXES = findobj(allchild(handle),'tag','AXES');
-    FIDSDISPLAY.XSLIDER = findobj(allchild(handle),'tag','SLIDERX');
-    FIDSDISPLAY.YSLIDER = findobj(allchild(handle),'tag','SLIDERY');
-    
+    FIDSDISPLAY.XWIN = [median([0 FIDSDISPLAY.XLIM]) median([3000/SCRIPTDATA.SAMPLEFREQ FIDSDISPLAY.XLIM])];
+    %    else
+    %FIDSDISPLAY.XWIN = [median([FIDSDISPLAY.XWIN(1) FIDSDISPLAY.XLIM]) median([FIDSDISPLAY.XWIN(2) FIDSDISPLAY.XLIM])];
+    %end
 
-    groups = SCRIPTDATA.DISPLAYGROUPF;
-    numgroups = length(groups);
-    
-    FIDSDISPLAY.NAME ={};
-    FIDSDISPLAY.GROUPNAME = {};
-    FIDSDISPLAY.GROUP = [];
-    FIDSDISPLAY.COLORLIST = {[1 0 0],[0 0.7 0],[0 0 1],[0.5 0 0],[0 0.3 0],[0 0 0.5],[1 0.3 0.3],[0.3 0.7 0.3],[0.3 0.3 1],[0.75 0 0],[0 0.45 0],[0 0 0.75]};
-    
-    if FIDSDISPLAY.MODE == 1
-        SCRIPTDATA.DISPLAYTYPEF1 = SCRIPTDATA.DISPLAYTYPEF;
-    else
-        SCRIPTDATA.DISPLAYTYPEF2 = SCRIPTDATA.DISPLAYTYPEF;
-    end    
-    
-    switch SCRIPTDATA.DISPLAYTYPEF
-        case 1   % show global RMS
[…]
