+++ SOURCE/GUI_Scripts/sliceDisplay.m
@@ -230,7 +230,7 @@ end
 
 
 function SetupDisplay(handle)
-
+    
 pointer = handle.Pointer;
 handle.Pointer = 'watch';
 global TS SCRIPTDATA SLICEDISPLAY;
@@ -260,6 +260,7 @@ numgroups = length(groups);
 SLICEDISPLAY.NAME ={};
 SLICEDISPLAY.GROUPNAME = {};
 SLICEDISPLAY.GROUP = [];
+
 
 switch SCRIPTDATA.DISPLAYTYPE
     case 1
@@ -305,17 +306,26 @@ switch SCRIPTDATA.DISPLAYTYPE
         for p=1:length(groups)
             SLICEDISPLAY.GROUPNAME{p} = [SCRIPTDATA.GROUPNAME{SCRIPTDATA.CURRENTRUNGROUP}{groups(p)}]; 
         end 
+
         SLICEDISPLAY.SIGNAL = TS{tsindex}.potvals(ch,:);
         SLICEDISPLAY.LEADINFO = TS{tsindex}.leadinfo(ch);
+         
 end
+
 
 switch SCRIPTDATA.DISPLAYSCALING
     case 1
-        k = max(abs(SLICEDISPLAY.SIGNAL),[],2);
-        [m,~] = size(SLICEDISPLAY.SIGNAL);
-        k(k==0) = 1;
-        s = sparse(1:m,1:m,1./k,m,m);
-        SLICEDISPLAY.SIGNAL = s*SLICEDISPLAY.SIGNAL;
+%         k = max(abs(SLICEDISPLAY.SIGNAL),[],2);
+%         [m,~] = size(SLICEDISPLAY.SIGNAL);
+%         k(k==0) = 1;
+%         s = sparse(1:m,1:m,1./k,m,m);
+%         SLICEDISPLAY.SIGNAL = s*SLICEDISPLAY.SIGNAL;
+            
+        SLICEDISPLAY.SIGNAL=SLICEDISPLAY.SIGNAL-min(SLICEDISPLAY.SIGNAL,[],2);
+
+        for p=1:size(SLICEDISPLAY.SIGNAL,1)
+            SLICEDISPLAY.SIGNAL(p,:)=SLICEDISPLAY.SIGNAL(p,:)/max(SLICEDISPLAY.SIGNAL(p,:));
+        end
     case 2
         k = max(abs(SLICEDISPLAY.SIGNAL(:)));
         [m,~] = size(SLICEDISPLAY.SIGNAL);
@@ -333,6 +343,7 @@ switch SCRIPTDATA.DISPLAYSCALING
         s = sparse(1:m,1:m,1./k,m,m);
         SLICEDISPLAY.SIGNAL = s*SLICEDISPLAY.SIGNAL;
 end
+
 
 if SCRIPTDATA.DISPLAYTYPE == 3
     SLICEDISPLAY.SIGNAL = 0.5*SLICEDISPLAY.SIGNAL+0.5;
@@ -358,7 +369,6 @@ switch SCRIPTDATA.DISPLAYOFFSET
 end
 
 SLICEDISPLAY.COLORLIST = {[1 0 0],[0 0.7 0],[0 0 1],[0.5 0 0],[0 0.3 0],[0 0 0.5],[1 0.3 0.3],[0.3 0.7 0.3],[0.3 0.3 1],[0.75 0 0],[0 0.45 0],[0 0 0.75]};
-
 handle.Pointer = pointer;
 
     
@@ -466,7 +476,9 @@ for p=chstart:chend
             color = [0.35 0.35 0.35];
         end
     end
+    
 
+    
[…]
