+++ SOURCE/Fiducializing/autoProcessSignal.m
@@ -49,7 +49,6 @@ idxs=round(linspace(1,length(leadsOfAllGroups),nToBeFiducialised));
 
 AUTOPROCESSING.leadsToAutoprocess=leadsOfAllGroups(idxs);
 
-
 %%%% get info from  already processed beat
 AUTOPROCESSING.bsk=TS{SCRIPTDATA.CURRENTTS}.selframes(1);    % "beat start kernel"
 AUTOPROCESSING.bek=TS{SCRIPTDATA.CURRENTTS}.selframes(2);  %beat end kernel

[…]
