+++ SOURCE/Fiducializing/findFid.m
@@ -37,9 +37,6 @@ function [globFid, indivFids, variance] = findFid(windows,kernels)
 
 
 nLeads=size(kernels,1);
-indivFids=zeros(nLeads,1);
-indivXcorr=zeros(nLeads,1);
-
 length_kernel=size(kernels,2);
 lagshift=0;
 numlags=size(windows,2)-length_kernel+1;   %only the lags with "no overlapping"
