### 2017-11-02 43feb77
fixed issue that in rare cases found wrong beats
