### 2017-11-03 f2f23bb
added USE_RMS option for autofiducialiazing
