+++ SOURCE/PFEIFER.m
@@ -2009,7 +2009,6 @@ function defaultsettings=getDefaultSettings
                     'ACTDEG',3,'integer',...
                     'RECWIN',7,'integer',...
                     'RECDEG',3,'integer',...
-                    'TRESHOLD_VAR', 50,'integer',...                %autofiducializing stuff
                     'FILTER_SELECTION',1,'toolsdropdownmenu',...    % tools options
                     'BASELINE_SELECTION',1,'toolsdropdownmenu',...
                     'ACT_SELECTION',1,'toolsdropdownmenu',...
@@ -2025,7 +2024,9 @@ function defaultsettings=getDefaultSettings
                     'ACCURACY', 0.9, 'double',...
                     'FIDSKERNELLENGTH',10,'integer',...
                     'WINDOW_WIDTH', 20, 'integer',...
-                    'NTOBEFIDUCIALISED', 10, 'integer'
+                    'NTOBEFIDUCIALISED', 10, 'integer',...
+                    'TRESHOLD_VAR', 50,'integer',...     
+                    'USE_RMS',1,'bool'
             };
 end
 
