### 2017-11-06 a252f04
added more autofiducializing options
