+++ SOURCE/PFEIFER.m
@@ -2026,7 +2026,8 @@ function defaultsettings=getDefaultSettings
                     'WINDOW_WIDTH', 20, 'integer',...
                     'NTOBEFIDUCIALISED', 10, 'integer',...
                     'TRESHOLD_VAR', 50,'integer',...     
-                    'USE_RMS',1,'bool'
+                    'USE_RMS',1,'bool',...
+                    'LEADS_FOR_AUTOFIDUCIALIZING',[],'vector'
             };
 end
 
