### 2017-11-13 027a1be
started working on entangling of autofiduciaolizng
