+++ SOURCE/Fiducializing/findFid.m
@@ -59,3 +59,4 @@ variance=var(indivFids);
 
 
 
+
