+++ SOURCE/Fiducializing/getBeatEnvelopes.m
@@ -1,0 +1,322 @@
+% MIT License
+% 
+% Copyright (c) 2017 The Scientific Computing and Imaging Institute
+% 
+% Permission is hereby granted, free of charge, to any person obtaining a copy
+% of this software and associated documentation files (the "Software"), to deal
+% in the Software without restriction, including without limitation the rights
+% to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+% copies of the Software, and to permit persons to whom the Software is
+% furnished to do so, subject to the following conditions:
+% 
+% The above copyright notice and this permission notice shall be included in all
+% copies or substantial portions of the Software.
+% 
+% THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+% IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+% FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+% AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+% LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+% OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+% SOFTWARE.
+
+
+
+
+
+function beatEnvelopes=getBeatEnvelopes(signal, kernel, accuracy)
+% finds kernel in signal.  retuns beatEnvelopes={[s_1,e_1], [s_2,e_2], ..[s_n,e_n]} so that 
+%  m=signal(s_i:e_i) matches kernel for all i, "matches" means:  xcorr(signal(s_i:e_i),kernel,0,'coeff') has a maximum peak value (at least accuracy)  (and matches dont overlap)
+% it also:
+% - sorts the beatEnvelopes in order they appear in signal (s1 < s2 < .. < si)
+global AUTOPROCESSING
+
+%%%% hardcoded:
+stepSize  = 20;
+accuracyForFirstEstimate = 0.7;
+
+
+%%%% set up stuff
+sigLength = length(signal);
+kernelLength = length(kernel);
+estNumBeats = round(sigLength/330);    % very rough estimation of numBeats in signal.. a beat is usually 330 frames long
+
+%%%% get stepLags: the evenly spaced lags (with distance stepSize) to go through. First possible lag would be 0! ("lag frame")
+stStepLag = 0.5 * stepSize;   % first lag
+endStepLag = sigLength  - kernelLength - 0.5 * stepSize; % last lag
+nStepLags = ceil((endStepLag - stStepLag)/stepSize) + 1;
+stepLags = round(linspace(stStepLag, endStepLag, nStepLags));
+
+
+%%%% get stepXC - the corresponding correlations for each stepLag
+count = 1;
+stepXCs = zeros(1,nStepLags);
+for lag = stepLags
+    stepXCs(count)=xcorr(signal(lag:lag+kernelLength-1), kernel,0,'coef');
+    count = count +1;
+end
+
+%%%% get the peakLags and corresponding peakXCs,  the correlations values and corresponding lags with high peak values
+estimatedPeakLags = zeros(1,estNumBeats);
+estimatedPeakXCVals = estimatedPeakLags;
+
+count=1;
+pval = 1;  %peak value
+while pval > accuracyForFirstEstimate
+    %get max val and corresponding lag
[…]
