### 2017-11-14 2bfc69e
made autofiducializing modular
