+++ SOURCE/Fiducializing/autoProcessSignal.m
@@ -31,28 +31,30 @@ function success = autoProcessSignal()
 clear global AUTOPROCESSING  % just in case, so previous stuff doesnt mess anything up
 global TS SCRIPTDATA AUTOPROCESSING
 
-unslicedDataIndex=SCRIPTDATA.unslicedDataIndex;
+unslicedDataIndex = SCRIPTDATA.unslicedDataIndex;
+templateFids = TS{SCRIPTDATA.CURRENTTS}.fids;
+templateBeatEnvelope = [TS{SCRIPTDATA.CURRENTTS}.selframes(1), TS{SCRIPTDATA.CURRENTTS}.selframes(2)];
+badLeads = (find(TS{SCRIPTDATA.unslicedDataIndex}.leadinfo > 0))';
 
 
-%%%% get the leadsToAutoprocess
-AUTOPROCESSING.leadsToAutoprocess = getLeadsToAutoprocess;
 
-%%%% get info from  already processed beat
-AUTOPROCESSING.bsk=TS{SCRIPTDATA.CURRENTTS}.selframes(1);    % "beat start kernel"
-AUTOPROCESSING.bek=TS{SCRIPTDATA.CURRENTTS}.selframes(2);  %beat end kernel
-AUTOPROCESSING.oriFids=TS{SCRIPTDATA.CURRENTTS}.fids;
+%%%%%% set up the settings structure
+settings.fidsKernelLength = SCRIPTDATA.FIDSKERNELLENGTH;
+settings.window_width = SCRIPTDATA.WINDOW_WIDTH;
+settings.nToBeFiducialized = SCRIPTDATA.NTOBEFIDUCIALISED;
+settings.leadsOfAllGroups = [SCRIPTDATA.GROUPLEADS{SCRIPTDATA.CURRENTRUNGROUP}{:}];
+settings.demandedLeads = SCRIPTDATA.LEADS_FOR_AUTOFIDUCIALIZING;
+settings.accuracy = SCRIPTDATA.ACCURACY;
+% TO DO:  USE_RMS ??
 
-
-%%%% get signal, the RMS needed to find beats
-leadsOfAllGroups=[SCRIPTDATA.GROUPLEADS{SCRIPTDATA.CURRENTRUNGROUP}{:}];
-badLeads=find(TS{SCRIPTDATA.unslicedDataIndex}.leadinfo > 0);
-leadsOfAllGroups = setdiff(leadsOfAllGroups, badLeads);
-signal = preprocessPotvals(TS{unslicedDataIndex}.potvals(leadsOfAllGroups,:));   % make signal out of leadsOfAllGroups
+SCRIPTDATA.DoIndivFids = 0;
+% TO DO: implement this
 
 
-%%%% find allFids based on oriFids and signal
-[AUTOPROCESSING.allFids, success]=findAllFids(TS{unslicedDataIndex}.potvals(AUTOPROCESSING.leadsToAutoprocess,:),signal);
+[AUTOPROCESSING.beats, AUTOPROCESSING.allFids, info, success] = getBeatsAndFids(TS{unslicedDataIndex}.potvals, templateBeatEnvelope, templateFids, badLeads, settings, SCRIPTDATA.DoIndivFids);
 if ~success, return, end
+
+AUTOPROCESSING.leadsToAutofiducialize = info.leadsToAutofiducialize;
 
 %%%% find AUTOPROCESSING.faultyBeatsIndeces and AUTOPROCESSING.faultyBeatInfo
 getFaultyBeats;
@@ -531,39 +533,3 @@ end
 
 
 
-function leadsToAutoprocess = getLeadsToAutoprocess
-% get leadsToAutoprocess, the leads to find fiducials for and plot.  Only these leads will be used to compute the global fids
[…]

[…]
