### 2017-11-14 2e97ee4
implemented USE_RMS for modular autofiducializing
