+++ SOURCE/Fiducializing/findAllFids.m
@@ -1,321 +1,0 @@
-% MIT License
-% 
-% Copyright (c) 2017 The Scientific Computing and Imaging Institute
-% 
-% Permission is hereby granted, free of charge, to any person obtaining a copy
-% of this software and associated documentation files (the "Software"), to deal
-% in the Software without restriction, including without limitation the rights
-% to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-% copies of the Software, and to permit persons to whom the Software is
-% furnished to do so, subject to the following conditions:
-% 
-% The above copyright notice and this permission notice shall be included in all
-% copies or substantial portions of the Software.
-% 
-% THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-% IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-% FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-% AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-% LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-% OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-% SOFTWARE.
-
-
-
-
-
-function [allFidsGlFr, success] = findAllFids(potvals,signal)
-% returns allFids = {fidsBeat1, fidsBeat2, .... , FidsLastBeat}
-% each fidsBeatN is a numFids - array - struct 'fids' with the fields: 'type' (the fiducial type) and  'value' (the values of the fid in global frame) 
-% in short: fidsBeatN is just like the 'fids' struct as it is saved in ts,
-% but in the global frame!!
-% Fids are determined based on 'oriFids', die fids done by the user.
-
-allFidsGlFr = 'dummyValue'; % to make sure this function has a value to return, even if function returns earlier as planned due to error
-%%%%% get paramters from SCRIPTDATA
-global SCRIPTDATA AUTOPROCESSING 
-accuracy=SCRIPTDATA.ACCURACY;  % abort condition5
-fidsKernelLength=SCRIPTDATA.FIDSKERNELLENGTH;  % the kernel indices will be from fidsValue-fidsKernelLength  until fidsValue+fidsKernelLength
-% reminder: it is kernel_idx=fid_start-fidsKernelLength:fid_start+fidsKernelLength
-window_width=SCRIPTDATA.WINDOW_WIDTH;   % dont search complete beat, but only a window with width window_width,
-% ws=bs+loc_fidsValues(fidNumber)-window_width;  
-% we=bs+loc_fidsValues(fidNumber)+window_width;
-totalKernelLength = 2*fidsKernelLength +1;   % the length of a kernel
-
-
-
-
-
-%%%% add the RMS to potvals, if USE_RMS is checked
-
-if SCRIPTDATA.USE_RMS
-    potvals(end+1,:) = signal;
-end
-
-
-
-
-%%%% beat kernel
-bsk=AUTOPROCESSING.bsk;   %start and end of beat
-bek=AUTOPROCESSING.bek;
-
-
-%%%% clear any nonglobal fids from oriFids
-oriFids=AUTOPROCESSING.oriFids;
-toBeCleared=[];
-for p=1:length(oriFids)
-    if length(oriFids(p).value)~=1   % if not single global value
-        toBeCleared=[toBeCleared p];
-    end
-end
-oriFids(toBeCleared)=[];
-
-
-
[…]
