+++ SOURCE/Fiducializing/oldVersion_autoProcessSignal.m
@@ -1,569 +1,0 @@
-% MIT License
-% 
-% Copyright (c) 2017 The Scientific Computing and Imaging Institute
-% 
-% Permission is hereby granted, free of charge, to any person obtaining a copy
-% of this software and associated documentation files (the "Software"), to deal
-% in the Software without restriction, including without limitation the rights
-% to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-% copies of the Software, and to permit persons to whom the Software is
-% furnished to do so, subject to the following conditions:
-% 
-% The above copyright notice and this permission notice shall be included in all
-% copies or substantial portions of the Software.
-% 
-% THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-% IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-% FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-% AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-% LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-% OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-% SOFTWARE.
-
-
-
-function success = oldVersion_autoProcessSignal()
-% do all the autoprocessing.  Use the fiducials in the fiducialed beat to find all other beats of 
-% that run and fiducialise those beats, too. Handle & save the autoprocessed beats just like the 
-% beat done by the user.
-
-%set up globals
-clear global AUTOPROCESSING  % just in case, so previous stuff doesnt mess anything up
-global TS SCRIPTDATA AUTOPROCESSING
-
-unslicedDataIndex=SCRIPTDATA.unslicedDataIndex;
-
-
-%%%% get the leadsToAutoprocess
-AUTOPROCESSING.leadsToAutoprocess = getLeadsToAutoprocess;
-
-%%%% get info from  already processed beat
-AUTOPROCESSING.bsk=TS{SCRIPTDATA.CURRENTTS}.selframes(1);    % "beat start kernel"
-AUTOPROCESSING.bek=TS{SCRIPTDATA.CURRENTTS}.selframes(2);  %beat end kernel
-AUTOPROCESSING.oriFids=TS{SCRIPTDATA.CURRENTTS}.fids;
-
-
-%%%% get signal, the RMS needed to find beats
-leadsOfAllGroups=[SCRIPTDATA.GROUPLEADS{SCRIPTDATA.CURRENTRUNGROUP}{:}];
-badLeads=find(TS{SCRIPTDATA.unslicedDataIndex}.leadinfo > 0);
-leadsOfAllGroups = setdiff(leadsOfAllGroups, badLeads);
-signal = preprocessPotvals(TS{unslicedDataIndex}.potvals(leadsOfAllGroups,:));   % make signal out of leadsOfAllGroups
-
-
-%%%% find allFids based on oriFids and signal
-[AUTOPROCESSING.allFids, success]=findAllFids(TS{unslicedDataIndex}.potvals(AUTOPROCESSING.leadsToAutoprocess,:),signal);
-if ~success, return, end
-
-%%%% find AUTOPROCESSING.faultyBeatsIndeces and AUTOPROCESSING.faultyBeatInfo
-getFaultyBeats;
-
-
[…]
