+++ SOURCE/PFEIFER.m
@@ -2029,7 +2029,7 @@ function defaultsettings=getDefaultSettings
                     'WINDOW_WIDTH', 20, 'integer',...
                     'NTOBEFIDUCIALISED', 10, 'integer',...
                     'TRESHOLD_VAR', 50,'integer',...     
-                    'USE_RMS',1,'bool',...
+                    'USE_RMS',0,'bool',...
                     'LEADS_FOR_AUTOFIDUCIALIZING',[],'vector'
             };
 end
