### 2017-11-17 #446 @bzenger
Hi @arfon,

Is there anything else we need to do to get an editor assigned?

Thanks
