### 2017-11-17 #446 @arfon
> Is there anything else we need to do to get an editor assigned?

Not right now. We need one of the JOSS editors to volunteer to edit this submission (which they've not yet done so).

Sorry for the delay.
