### 2017-11-17 #446 @nirum
Hi, I don’t have the bandwidth right now to give this a thorough review. Apologies! - Niru

> On Nov 17, 2017, at 9:01 AM, Arfon Smith <notifications@github.com> wrote:
>
> Is there anything else we need to do to get an editor assigned?
>
> Not right now. We need one of the JOSS editors to volunteer to edit this submission (which they've not yet done so).
>
> Sorry for the delay.
>
