### 2017-11-20 #446 @arfon
@Kevin-Mattheus-Moerman @cMadan - would either of you be willing to edit this submission for JOSS? I realize it's a little out of any of our expertise areas but we could look at getting an additional reviewer to help here.
