### 2017-11-22 afcde3b
improved autofiducializing
