+++ SOURCE/Fiducializing/getBeatsAndFids.m
@@ -21,11 +21,7 @@
 % SOFTWARE.
 
 
-
-
-
-
-function [beatEnvelopes, allFidsAbsFr, info, success] = getBeatsAndFids(unslicedComplPotvals,templateBeatEnvelope, templateFids, badLeads, settings, DoIndivFids)
+function [allBeatEnvelopes, allFidsAbsFr, info, success] = getBeatsAndFids(unslicedComplPotvals,templateBeatEnvelope, templateFids, badLeads, settings)
 % INPUTS:
 %   - unslicedComplPotvals: the complete (all leads) and unsliced (not sliced into beats) potvals where want to find teh beats and fiducials
 %     any preprocessing like temporal filtering should already be done
@@ -40,20 +36,76 @@ function [beatEnvelopes, allFidsAbsFr, info, success] = getBeatsAndFids(unsliced
 %           - accuracy
 %           - fidsKernelLength
 %           - window_width
-%   - DoIndivFids: a bool indicating if you want individual fids or only global fids
+%           - DoIndivFids: a bool indicating if you want individual fids or only global fids
+%           - autoUpdateKernels: a bool indicating if you want to update the template once every nBeatsBeforeUpdating
+%           - nBeatsToAvrgOver:  an integer indicating over how many beats you want to average to get a new updated fiducial kernel.
+%             This field is not needed if autoUpdateKernels == 0.
+%           - nBeatsBeforeUpdating: an integer indicating how often (after how many beats) you want to update the fiducial kernel
+%             This field is not needed if autoUpdateKernels == 0.
+
 
 
 
 %%%% initialize outputs and some other stuff
-beatEnvelopes = {};
-allLocFrFids = {};
-success = 0;
+allBeatEnvelopes = {};
+allFidsAbsFr = {};
+info = [];
+
 nLeads = size(unslicedComplPotvals,1);
 nFrames = size(unslicedComplPotvals,2);
 bsk = templateBeatEnvelope(1);
 bek = templateBeatEnvelope(2);
 
 
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%%%%%%%%%%%%%%%%%%%% get the (first) beat envelopes %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%%%% get the RMS of unslicedPotvals
+leadsForRMS = setdiff(settings.leadsOfAllGroups, badLeads);
+rmsSignal = preprocessPotvals(unslicedComplPotvals(leadsForRMS,:));
+
+%%%% get the (first) beatEnvelopes
+if settings.autoUpdateKernels
+    lengthOfSearchArea = settings.nBeatsBeforeUpdating * 2000;
+    if length(rmsSignal) < lengthOfSearchArea
+        searchArea = 1:length(rmsSignal);
+    else
+        searchArea = 1:lengthOfSearchArea;
+    end
+    
+    allBeatEnvelopes = getBeatEnvelopes(rmsSignal(searchArea), rmsSignal(bsk:bek), settings.accuracy,bsk);
+    % find oriBeatIdx, the index of the template beat
+    oriBeatIdx = [];
+    for beatCount=1:length(allBeatEnvelopes)    
+        if (abs(allBeatEnvelopes{beatCount}(1)-bsk)) < 3  % if found beat "close enough" to original Beat 
[…]
