### 2017-11-23 #446 @CristianaCorsi
My full name is Cristiana Corsi and my istitutional page is
https://www.unibo.it/sitoweb/cristiana.corsi3/en

Kind regards,
Cristiana Corsi

  Il 22/11/2017 22.04, Kevin Mattheus Moerman ha scritto:
>
> Thanks. @CristianaCorsi <https://github.com/cristianacorsi> you have
> been suggested as potential reviewer. If you are interested in
> reviewing can you please share​ your full name and perhaps a link to a
> professional (e.g. institutional) page? I ask this for verification
> and to learn about your affiliation/status/experience since your
> GitHub profile seems brand new and is currently empty.
>
