### 2017-11-27 #472 @whedon
[REVIEW]: PFEIFER: Preprocessing Framework of Electrograms Intermittently Fiducialized from Experimental Recordings
**Submitting author:** @bzenger (<a href="http://orcid.org/0000-0002-0039-9184">Brian Zenger</a>)
**Repository:** <a href="https://github.com/SCIInstitute/PFEIFER" target ="_blank">https://github.com/SCIInstitute/PFEIFER</a>
**Version:** v1.0.0

I suggest to use in the Documentation, figures  that describe exactly the windows that open when you run PFEIFER. For example, in Chapter 6, figure 6.1 shows a window slightly different from the DATA_ORGANIZATION window that opens when you run PFEIFER. In the window each section has a specific name and the BEAT PROPAGATION section includes six fields, not four as reported in the figure (Window Length is called Window Width; the fields #Beats before updating and # Beats for updating are missing as well as the flags “Include RMS for Autofiducializing”, “Local Fiducials”, “Update Templates”; and the button !Default Settings!). The design of the window is also different from the one reported in figure 6.1.
I also suggest to include in the Documentation the link to the tutorials available at: http://www.sci.utah.edu/software/pfeifer.html
