### 2017-11-27 #472 @Kevin-Mattheus-Moerman
@CristianaCorsi this is where the review takes place. The review is guided by the tick boxes at the top of this issue which you are able to tick. See our website and [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) for more information. If you have comments or issues you want to report for the authors you can leave comments here. You can also post issues on the project repo and link to them here. Thanks you. Let me know if I can help with anything.
