### 2017-11-27 #472 @Kevin-Mattheus-Moerman
@bzenger The initial paper generation did not complete. I am unsure about the error but it looks like it cannot find your paper file. Could you please rename the paper folder and paper file to `paper` and `paper.md` respectively? Thanks
