+++ SOURCE/Fiducializing/plotAutoProcFids.m
@@ -68,7 +68,13 @@ slidery=findobj(allchild(fig),'tag','SLIDERY');
 addlistener(sliderx,'ContinuousValueChange',@UpdateSlider);
 addlistener(slidery,'ContinuousValueChange',@UpdateSlider);
 
-AUTOPROCESSING.DISPLAYTYPEA = 1;
+%%%% set up fiducial selection
+AUTOPROCESSING.DISPLAYTYPEA = 1;  % DISPLAYTYPEA is either 1 or 3 to show either global RMS or individual leads
+AUTOPROCESSING.SELFIDS = 1;       % is either 1 or 3 letting you edit either the local or the global fiducials
+set(findobj(allchild(fig),'tag','FIDSGLOBAL'),'value',1);
+set(findobj(allchild(fig),'tag','FIDSLOCAL'),'value',0);
+
+
 
 %%%% init AUTOPROCESSING
 AUTOPROCESSING.CurrentEventIdx=[];
@@ -149,8 +155,6 @@ end
 beatSelectPopupObj.String = selectionChoises;
 
 
-
-
 function setupDisplay(fig)
 %      no plotting, but everything else with axes, particualrely:
 %         - sets up some start values for xlim, ylim, sets up axes and slider handles
@@ -195,17 +199,17 @@ AUTOPROCESSING.COLORLIST = {[1 0 0],[0 0.7 0],[0 0 1],[0.5 0 0],[0 0.3 0],[0 0 0
 if ~isfield(SCRIPTDATA,'DISPLAYTYPEA')
     SCRIPTDATA.DISPLAYTYPEA=1;
 end
-%%%% set up signals for global RMS, GROUP RMS or individual RMS
+%%%% set up signals for global RMS, GROUP RMS or individual leads
 switch SCRIPTDATA.DISPLAYTYPEA
     case 1   % show global RMS
-        ch  = []; 
+        allLeadToPlots  = []; 
         for p=groups 
             leads = SCRIPTDATA.GROUPLEADS{SCRIPTDATA.CURRENTRUNGROUP}{p};
             index = TS{tsindex}.leadinfo(leads)==0;  % index of only the 'good' leads, filter out badleads
-            ch = [ch leads(index)];   % ch is leads only of the leads of the groubs selected, not of all leads
+            allLeadToPlots = [allLeadToPlots leads(index)];   % ch is leads only of the leads of the groubs selected, not of all leads
         end
         
-        AUTOPROCESSING.SIGNAL = sqrt(mean(TS{tsindex}.potvals(ch,:).^2));
+        AUTOPROCESSING.SIGNAL = sqrt(mean(TS{tsindex}.potvals(allLeadToPlots,:).^2));
         AUTOPROCESSING.SIGNAL = AUTOPROCESSING.SIGNAL-min(AUTOPROCESSING.SIGNAL);
         AUTOPROCESSING.LEADINFO = 0;
         AUTOPROCESSING.GROUP = 1;
@@ -213,8 +217,21 @@ switch SCRIPTDATA.DISPLAYTYPEA
         AUTOPROCESSING.LEADGROUP = 0;
         AUTOPROCESSING.NAME = {'Global RMS'};
         AUTOPROCESSING.GROUPNAME = {'Global RMS'};
+        
+        
+        %%%% set the indiv/global fids buttons,  disable/enable the right ones
+        if AUTOPROCESSING.SELFIDS > 1
+            AUTOPROCESSING.SELFIDS = 1;
+            set(findobj(allchild(fig),'tag','FIDSGLOBAL'),'value',1);
+            set(findobj(allchild(fig),'tag','FIDSLOCAL'),'value',0);
+        end
[…]
