### 2017-11-28 #472 @Kevin-Mattheus-Moerman
@bzenger that seems to have worked. Can  you check if the affiliations are complete/correct in [your article](https://github.com/openjournals/joss-papers/blob/joss.00472/joss.00472/10.21105.joss.00472.pdf)?
