### 2017-11-29 #472 @arfon
@bzenger - sorry for the unhelpful errors from @whedon. I've made some fixes to your paper YAML header that were obvious problems: https://github.com/SCIInstitute/PFEIFER/pull/24
