+++ SOURCE/Fiducializing/getBeatsAndFids.m
@@ -103,7 +103,6 @@ else
     nBeats = length(allBeatEnvelopes);
 end
 
-
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%%%%%%%%%%%%%%%%%%% get unslicedReducedPotvals, the leads to work with  %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
@@ -135,7 +134,8 @@ end
 
 %%%% baselineCorrect the beats we have so far
 unslicedReducedPotvals = baselineCorrectBeats(unslicedReducedPotvals,allBeatEnvelopes);
-
+% set values before first beat to zero  (this makes things easier later, since we don't have to worry anymore if window is out of beat
+unslicedReducedPotvals(:,1:allBeatEnvelopes{1}(1)) = zeros(nLeadsToAutofiducialize,allBeatEnvelopes{1}(1));
 
 
 
@@ -164,11 +164,8 @@ blcs = bs + relFrOriFidsValues(fidTypes == 2) - 1 - 40;
 blce = bs + relFrOriFidsValues(fidTypes == 4) - 1 + 50;
 if blcs < 1
     blcs = 1;
-    disp('WARNING: qrs start is pretty close to beat start. This can cause in bad fiducial detection.')
+    disp('WARNING: qrs start is pretty close to beat start. This can result in bad fiducial detection.')
 end
-
-
-
 unslicedReducedPotvals(:,blcs:blce) = baselineCorrection(unslicedReducedPotvals(:,blcs:blce),1, blce-blcs,5);
 
 
@@ -200,7 +197,6 @@ end
 
 %%%% set up the kernels
 totalKernelLength = 2 * settings.fidsKernelLength +1;
-
 fidKernel = zeros(nLeadsToAutofiducialize, totalKernelLength, nFids);
 for fidIdx = 1:nFids
     fidKernel(:,:,fidIdx) = unslicedReducedPotvals(:,fsk(fidIdx):fek(fidIdx));
@@ -209,10 +205,10 @@ end
 
 
 
-%%%% for testing only
+% %%% for testing only
 % clear global xxx
 % global xxx
-% xxx.pv = unslicedReducedPotvals;
+% xxx.pv3 = unslicedReducedPotvals;
 % xxx.fsk = fsk;
 % xxx.fek = fek;
 % xxx.fidKernel = fidKernel;
@@ -260,16 +256,27 @@ while beatCount <= nBeats + 1  % as long as there are still beat envelopes with 
             for p=1:length(newBeatEnvelopes)
                 newBeatEnvelopes{p} = newBeatEnvelopes{p} + reference - 1;
             end
-
+            
+            % get rid of superfluous newBeatEnvelopes. We only want nBeatsBeforeUpdating new beats
             if length(newBeatEnvelopes) > settings.nBeatsBeforeUpdating
                 newBeatEnvelopes = newBeatEnvelopes(1:settings.nBeatsBeforeUpdating);
             end
-            nBeats = nBeats + length(newBeatEnvelopes);
-            allBeatEnvelopes(beatCount:nBeats) = newBeatEnvelopes;
             
             
             %%%% baselineCorrect the new found beats
             unslicedReducedPotvals = baselineCorrectBeats(unslicedReducedPotvals,newBeatEnvelopes);
+            %also set values between last old beat and first new beat to zero
+            if ~isempty(newBeatEnvelopes)
[…]
