### 2017-11-30 #472 @Kevin-Mattheus-Moerman
@arfon check the second to last reference in the pdf. The DOI seems to spill over the margin. Can that be fixed?
