### 2017-11-30 #472 @Kevin-Mattheus-Moerman
@bzenger Can you check the Good et al. reference.Please add a link to the DOI (I think it should probably be: 10.23919/CIC.2016.7868928 see [here](http://ieeexplore.ieee.org/document/7868928/)).
