### 2017-11-30 #472 @arfon
> @arfon check the second to last reference in the pdf. The DOI seems to spill over the margin. Can that be fixed?

There must be right? It's beyond my LaTeX super-powers though I'm afraid. We had a go in https://github.com/openjournals/whedon/pull/14 but this didn't help.
