### 2017-11-30 #472 @Kevin-Mattheus-Moerman
@arfon I take it you've tried some of [these suggestions](https://tex.stackexchange.com/questions/3033/forcing-linebreaks-in-url)?
