### 2017-11-30 #472 @arfon
> @arfon I take it you've tried some of these suggestions?

Yes... and to no end I'm afraid.
