### 2017-12-01 #472 @Kevin-Mattheus-Moerman
@arfon. Perhaps we can fix this by always starting the DOI link on a new line.
