### 2017-12-01 #472 @arfon
> @arfon. Perhaps we can fix this by always starting the DOI link on a new line.

Perhaps. Not sure how we would do that though?
