### 2017-12-01 #472 @Kevin-Mattheus-Moerman
I've added some statements over at https://github.com/openjournals/whedon/pull/14.
