### 2017-12-08 #472 @Kevin-Mattheus-Moerman
@CristianaCorsi thanks for the update! Our review guidelines are [here](http://joss.theoj.org/about#reviewer_guidelines) and you can find the tick-boxes at the top of this issue. Thanks for your help reviewing!
