### 2018-01-22 #472 @CristianaCorsi
Hi Kevin.

I completed the review process.
I checked all the flags in the checklist and added some comments to the
Documentation section.
In my opinion the paper should be accepted.

Is the review process finalized or I need to upload some additional
files? Do I need to apply some different label to this issue?

Thank you!
Cristiana

Il 08/01/2018 17.22, Kevin Mattheus Moerman ha scritto:
>
> @bzenger <https://github.com/bzenger> @CristianaCorsi
> <https://github.com/cristianacorsi> Happy New Year! 🎉
> @CristianaCorsi <https://github.com/cristianacorsi> how are you
> getting on? Are you able to finalize the review process soon? Let me
> know if I can help with anything.
>
