### 2018-01-22 #472 @Kevin-Mattheus-Moerman
@ChristianaCorsi :tada: excellent thanks for completing the review process. We will handle things from here on. Can you and @bzenger confirm the comments to the documentation have been implemented? Can you link to an issue (if an issue was used)?
