### 2018-01-22 #472 @bzenger
Hi Kevin,

We just updated our documentation with some small changes. These were done over an email exchange not through any GitHub issue.

Dr. Corsi will email you with the confirmation.

Thanks!

Brian

> On Jan 22, 2018, at 10:11 AM, Kevin Mattheus Moerman <notifications@github.com> wrote:
>
> @christianacorsi 🎉 excellent thanks for completing the review process. We will handle things from here on. Can you and @bzenger <https://github.com/bzenger> confirm the comments to the documentation have been implemented? Can you link to an issue (if an issue was used)?
>
