### 2018-01-23 #472 @CristianaCorsi
All the comments to the documentation have been implemented.

Brian, could you please update the DOC folder in the submission web site
with this new file (now available on the server
http://sci.utah.edu/devbuilds/pfeifer_docs/PfeiferDocumentation.pdf)?
Thank you!

Cristiana

Il 22/01/2018 18.11, Kevin Mattheus Moerman ha scritto:
>
> @christianacorsi 🎉 excellent thanks for completing the review
> process. We will handle things from here on. Can you and @bzenger
> <https://github.com/bzenger> confirm the comments to the documentation
> have been implemented? Can you link to an issue (if an issue was used)?
>
