### 2018-01-24 #472 @Kevin-Mattheus-Moerman
Okay it looks like we are all set here! 🚀 🎉
@bzenger  At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? @arfon can then move forward with accepting the submission.
Thanks @CristianaCorsi for the review!
